refactor(analytics): migrate labelsFormatter to utils and update yAxis label formatting
