+++ src/pages/Analytics/Cards/TaskD/data.js
@@ -2,7 +2,8 @@
 
 import Highcharts from 'highcharts';
 
-import { groupYAxisLabelFormat, parsedData } from './parsed-files';
+import { labelsFormatter } from 'pages/Analytics/utils/utils';
+import { parsedData } from './parsed-files';
 
 export const seriesToShow = ['public', 'private'];
 const colorIndex = [3, 4];
@@ -45,56 +46,6 @@ const prepareXAxis = series => ({
     }, {}),
   ),
 });
-
-/**
-
- * @function prepareYAxisLabelFormat
- * @param {TaskDTypes.FileKey} label - Selected value from the filter.
- * @returns {TaskDTypes.YAxisLabelFormatValue} Formatted label based on the selected value.
- */
-export const prepareYAxisLabelFormat = label => groupYAxisLabelFormat[label];
-
-/**
-
- * @function labelsFormatter
- * @param {TaskDTypes.FileKey} selectedValue - Selected value from the filter.
- * @param {string} lng - Language code.
- * @returns {string} Formatted label based on the selected value and language.
- */
-export function labelsFormatter(selectedValue, lng) {
-  if (!selectedValue || !lng) {
-    throw new Error('Both selectedValue and lng are required.');
-  }
-
-  const labelFormat = prepareYAxisLabelFormat(selectedValue);
-  const localeBase = lng.split('-')[0] || 'en';
-  const suffixMap = {
-    sl: 'tis.',
-    en: 'k',
-  };
-
-  const { value } = this;
-  if (value == null) return '';
-
-  if (labelFormat === 'percent') {
-    return new Intl.NumberFormat(lng, {
-      style: 'percent',
-      minimumFractionDigits: 0,
-      maximumFractionDigits: 2,
-    }).format(value / 100);
-  }
-
-  const isOverMillion = value >= 1_000_000;
-  const suffix = isOverMillion ? suffixMap[localeBase] || 'k' : '';
-  const valueToFormat = isOverMillion ? value / 1_000 : value;
-
-  return `${new Intl.NumberFormat(lng, {
-    style: 'decimal',
-    minimumFractionDigits: 0,
-    maximumFractionDigits: 2,
-    useGrouping: 'auto',
-  }).format(valueToFormat)} ${suffix}`.trim();
-}
 
 const prepareYAxis = (filterState, lng, title) => ({
   title: {
