+++ src/pages/Analytics/Cards/TaskD/parsed-files.js
@@ -41,6 +41,7 @@ export const groupYAxisLabelFormat = Object.freeze({
   gpObseg: 'decimal',
   denDoseganjePovprecja: 'percent',
   denObseg: 'decimal',
+  taskSpecialLeft: 'decimal',
 });
 
 /**
